Rename & move UiTPASControllerProvider for labels for clarity
