+++ app/UiTPASService/UiTPASServiceLabelsControllerProvider.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Silex\UiTPAS;
+namespace CultuurNet\UDB3\Silex\UiTPASService;
 
 use Silex\Application;
 use Silex\ControllerCollection;
 use Silex\ControllerProviderInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
-final class UiTPASControllerProvider implements ControllerProviderInterface
+final class UiTPASServiceLabelsControllerProvider implements ControllerProviderInterface
 {
     public function connect(Application $app): ControllerCollection
     {
@@ -17,7 +17,7 @@ final class UiTPASControllerProvider implements ControllerProviderInterface
         $controllers = $app['controllers_factory'];
 
         $controllers->get(
-            '/labels',
+            '/',
             function (Application $app) {
                 return new JsonResponse(
                     $app['config']['uitpas']['labels']
